# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/non_conformances:
    get:
      tags:
        - NonConformances
      summary: List NonConformances
      description: Get a paginated list of NonConformances with optional filtering and sorting
      operationId: listNonConformances
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonConformanceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - NonConformances
      summary: Create NonConformance
      description: Create a new NonConformance
      operationId: createNonConformance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNonConformanceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonConformance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/quality_actions:
    get:
      tags:
        - QualityActions
      summary: List QualityActions
      description: Get a paginated list of QualityActions with optional filtering and sorting
      operationId: listQualityActions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityActionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QualityActions
      summary: Create QualityAction
      description: Create a new QualityAction
      operationId: createQualityAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQualityActionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    NonConformance:
      type: object
      required:
        - id
        - company_id
        - subject
        - severity
        - status
        - opened_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        source_inspection_id:
          nullable: true
          type: string
          format: uuid
        item_id:
          nullable: true
          type: string
          format: uuid
        severity:
          type: string
        status:
          type: string
        description:
          nullable: true
          type: string
          maxLength: 4000
        opened_at:
          type: string
          format: date-time
        closed_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateNonConformanceRequest:
      type: object
      required:
        - company_id
        - subject
        - severity
        - status
        - opened_at
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        source_inspection_id:
          nullable: true
          type: string
          format: uuid
        item_id:
          nullable: true
          type: string
          format: uuid
        severity:
          type: string
        status:
          type: string
        description:
          nullable: true
          type: string
          maxLength: 4000
        opened_at:
          type: string
          format: date-time
        closed_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    NonConformanceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NonConformance'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    QualityAction:
      type: object
      required:
        - id
        - company_id
        - non_conformance_id
        - action_type
        - status
        - description
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        non_conformance_id:
          type: string
          format: uuid
        action_type:
          type: string
        procedure_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        description:
          type: string
          maxLength: 2000
        due_date:
          nullable: true
          type: string
          format: date-time
        completed_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateQualityActionRequest:
      type: object
      required:
        - company_id
        - non_conformance_id
        - action_type
        - status
        - description
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        non_conformance_id:
          type: string
          format: uuid
        action_type:
          type: string
        procedure_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        description:
          type: string
          maxLength: 2000
        due_date:
          nullable: true
          type: string
          format: date-time
        completed_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    QualityActionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QualityAction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []