# src/presentation/http/non_conformance_handler.rs
//! NonConformance REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{NonConformanceService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateNonConformanceDto, UpdateNonConformanceDto, PatchNonConformanceDto, NonConformanceResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum NonConformanceError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for NonConformanceError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for NonConformanceError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "NONCONFORMANCE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "NONCONFORMANCE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "NONCONFORMANCE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "NONCONFORMANCE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for NonConformance.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /non_conformances | List with pagination |
/// | POST | /non_conformances | Create new |
/// | GET | /non_conformances/:id | Get by ID |
/// | PUT | /non_conformances/:id | Full update |
/// | PATCH | /non_conformances/:id | Partial update |
/// | DELETE | /non_conformances/:id | Soft delete |
/// | POST | /non_conformances/bulk | Bulk create |
/// | POST | /non_conformances/upsert | Upsert |
/// | GET | /non_conformances/trash | List deleted |
/// | POST | /non_conformances/:id/restore | Restore |
/// | DELETE | /non_conformances/empty | Empty trash |
/// | GET | /non_conformances/:id/deleted | Get deleted by ID |
/// | DELETE | /non_conformances/trash/:id | Permanent delete from trash |
/// | GET | /non_conformances/count | Count active entities |
/// | GET | /non_conformances/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(NonConformanceService::with_repository(repository));
/// let router = create_non_conformance_routes(service);
/// ```
pub fn create_non_conformance_routes(service: Arc<NonConformanceService>) -> Router {
    BackboneCrudHandler::<NonConformanceService, NonConformance, CreateNonConformanceDto, UpdateNonConformanceDto, NonConformanceResponseDto>::routes(
        service,
        "/non_conformances",
    )
}

/// Create Axum router with only the read (GET) endpoints for NonConformance.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_non_conformance_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_non_conformance_read_routes(service: Arc<NonConformanceService>) -> Router {
    BackboneCrudHandler::<NonConformanceService, NonConformance, CreateNonConformanceDto, UpdateNonConformanceDto, NonConformanceResponseDto>::read_routes(
        service,
        "/non_conformances",
    )
}

/// Create Axum router with only the write (mutation) endpoints for NonConformance.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_non_conformance_write_routes(service: Arc<NonConformanceService>) -> Router {
    BackboneCrudHandler::<NonConformanceService, NonConformance, CreateNonConformanceDto, UpdateNonConformanceDto, NonConformanceResponseDto>::write_routes(
        service,
        "/non_conformances",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_non_conformance_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<NonConformanceService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_non_conformance_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}